refactor(examples) use playwright testrunner
