+++ examples/next-prisma-starter/package.json
@@ -19,7 +19,7 @@
     "test-start": "start-server-and-test start 3000 test",
     "test": "run-s test:*",
     "test:unit": "jest",
-    "test:e2e": "jest --config jest.playwright.config.ts",
+    "test:e2e": "playwright test",
     "ts-node": "ts-node --compiler-options \"{\\\"module\\\":\\\"commonjs\\\"}\"",
     "postinstall": "prisma generate"
   },
@@ -47,6 +47,7 @@
     "zod": "^3.0.0"
   },
   "devDependencies": {
+    "@playwright/test": "^1.19.1",
     "@types/jest": "^27.0.1",
     "@types/node": "^17.0.10",
     "@types/react": "^17.0.20",
@@ -60,8 +61,6 @@
     "eslint-plugin-react": "^7.25.1",
     "eslint-plugin-react-hooks": "^4.2.0",
     "jest": "^27.1.0",
-    "jest-playwright": "^0.0.1",
-    "jest-playwright-preset": "^1.4.5",
     "npm-run-all": "^4.1.5",
     "playwright": "^1.14.1",
     "prettier": "^2.5.1",
