+++ examples/next-prisma-starter/tsconfig.json
@@ -20,7 +20,6 @@
     "jsx": "preserve",
     "types": [
       "@types/jest",
-      "jest-playwright-preset",
       "expect-playwright"
     ],
     "baseUrl": "./src",
